converted the options in a drop down menu.
